tidycells: Read Tabular Data from Diverse Sources and Easily Make Them Tidy

Provides utilities to read, cells from complex tabular data and heuristic detection based 'structural assignment' of those cells to a columnar or tidy format. Read functionality has the ability to read structured, partially structured or unstructured tabular data from various types of documents. The 'structural assignment' functionality has both supervised and unsupervised way of assigning cells data to columnar/tidy format. Multiple disconnected blocks of tables in a single sheet are also handled appropriately. These tools are suitable for unattended conversation of messy tables into a consumable format(usable for further analysis and data wrangling).

Get Started

To get started check out vignette("tidycells-intro").

Alternatively, You can check out A Brief Introduction for quick start (and useful information about how to utilize help for tidycells).

Main Functions

Here is a glimpse of main functions from this package. (There are other functionality available which are briefed in A Brief Introduction.)

as_cell_df

value_attribute_classify

analyze_cells

compose_cells

collate_columns

And a function to combine all these (above listed) functionality:

read_cells

See also